separated led control into library
