+++ include/PinConfig.h
@@ -4,6 +4,10 @@
 /* Pump Control */
 const int pumpControlPin = 9;
 
+/* RGB LED */
+const byte GreenLedPin = 32;
+const byte RedLedPin = 26;
+const byte BlueLedPin = 27;
 
 
 #endif
