+++ lib/LedControl/LedControlConstants.h
@@ -1,0 +1,16 @@
+#ifndef LED_CONTROL_CONSTANTS_H
+#define LED_CONTROL_CONSTANTS_H
+
+
+#define PWM1_Ch    0
+#define PWM1_Res   8
+#define PWM1_Freq  1000
+#define PWM2_Ch    1
+#define PWM2_Res   8
+#define PWM2_Freq  1000
+#define PWM3_Ch    2
+#define PWM3_Res   8
+#define PWM3_Freq  1000
+
+
+#endif
